+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/activities/variables/VariablesViewState.kt
@@ -8,4 +8,7 @@ data class VariablesViewState(
 ) {
     val isDraggingEnabled: Boolean
         get() = variables.size > 1
+
+    val isSortButtonEnabled
+        get() = variables.size > 1
 }